+++ src/Trading/Services/TradingService.php
@@ -9,7 +9,7 @@ namespace DTS\eBaySDK\Trading\Services;
 
 class TradingService extends \DTS\eBaySDK\Trading\Services\TradingBaseService
 {
-    const API_VERSION = '979';
+    const API_VERSION = '981';
 
     /**
      * @param array $config Configuration option values.