+++ src/Types/DecimalType.php
@@ -13,7 +13,7 @@ class DecimalType extends \DTS\eBaySDK\Types\BaseType
      */
     private static $propertyTypes = [
         'value' => [
-            'type' => 'integer',
+            'type' => 'integer|double',
             'repeatable' => false,
             'attribute' => false
         ]