+++ test/Types/DecimalTypeTest.php
@@ -25,5 +25,25 @@ class DecimalTypeTest extends \PHPUnit_Framework_TestCase
         $this->obj->value = 123;
         $this->assertEquals(123, $this->obj->value);
         $this->assertInternalType('integer', $this->obj->value);
+
+        $this->obj->value = -123;
+        $this->assertEquals(-123, $this->obj->value);
+        $this->assertInternalType('integer', $this->obj->value);
+
+        $this->obj->value = 123.00;
+        $this->assertEquals(123.00, $this->obj->value);
+        $this->assertInternalType('float', $this->obj->value);
+
+        $this->obj->value = -123.00;
+        $this->assertEquals(-123.00, $this->obj->value);
+        $this->assertInternalType('float', $this->obj->value);
+
+        $this->obj->value = 123.45;
+        $this->assertEquals(123.45, $this->obj->value);
+        $this->assertInternalType('float', $this->obj->value);
+
+        $this->obj->value = -123.45;
+        $this->assertEquals(-123.45, $this->obj->value);
+        $this->assertInternalType('float', $this->obj->value);
     }
 }