+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Unreleased
+
+### Breaking changes
+
+* Support Merchant Data API version 981.
+
 ## 6.0.0 - 2016-08-30
 
 ### Breaking changes