+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Support Merchant Data API version 981.
 * Support Shopping API version 981.
+* Support Trading API version 981.
 
 ## 6.0.0 - 2016-08-30